refactor: Subscribe to store and update its state
Added a local state to the App component and an input field, to let a user add a post, which for now is hardcoded to appear in the `redux` category.
Added a `subscribe` method to the store to update the local component state every time the store was modified by a user input, causing a re-render that would show that input on the screen.