refactor: Change initialState and reducer
Changed the initial state of the application following the Redux documentation that advices on giving a shape to the state using normalized data.
To avoid a deep nesting of objects referencing each other, it was used an approach that would reproduce relational databases tables, with each piece of data separated and external keys to reference different objects inside other objects.
A key named entity was added to contain the different object entities that would shape the application, with other data referred by different properties.