refactor: Add comment id to the related post
Added code to add the comment id, as a user comments a post, to the comments array in the related post.
Added the `addComment` action type to the post reducer to manage this job.
Studied the code example given in the Redux docs about Updating Normalized Data in the Structuring Reducers section to learn how to apply actions to different slices of the state.