feat: Add action to add a new post to the server
Wrote an action to be dispatched by the `thunk` middleware to add a new post to the server.
NOTE that code at this stage needs to be heavily refactored and cleaned. Commented out code has been left in place to serve as reminder of the steps taken to move from a synchronous application to its asynchronous version which could communicate with the server (this is after all a learning experience). This version only adds posts to the server and lists them in the `allPosts` array. All previous functionality will be added gradually later on.