feat: Install logger and compose enhancers
Installed logger to log to the console the dispatched actions. Used Redux middleware for this purpose, as it would otherwise break the rule for Redux reducers to be pure functions.
Added logger as the last middleware in chain to avoid it logging thunk and promise, instead of actual actions (as requested in the official documentation). Composed the enhancer middleware together with the Redux dev tools invocation.