refactor: Populate posts and add comments property
Changed code to make the post object be populated as the App component mounted and have the comments be selected from the server for each post item. Each post object would have a `comments` property key added, referencing an array with the ids of its related comments.
This way, as the App component mounted, the main view of the app could show all the available posts from the server and their porperties, without reference to a specific category.