refactor: Clean code and add comments
Refactored code to delete commented out code and add meaningful comments.
At this stage the store gets updated correctly from the server as the App component mounts, receiving the stored posts and related comments.
The app needs to be refactored to modify actions triggered from the `submit` methods on both posts and comments forms and is ready to add the React code to be completed.