refactored: Navigate with React Router <Link>
Modified code to make the Router work with Redux. Got inspiration and ideas on how to structure my code from the Redux official docs and the section `Usage with React Router` from the advanced tutorial. The official `Todos` example and the walkthrough explanation on how to connect it with React Router gave me the direction to head my project and to make effective use of the component.
Got solutions from Stackoverflow on how to define a path parameter in React Router v4, that differs from what's written in the official documentation (still at version 2) and on how to render a <Route> exclusively using <Switch>.