feat: Add a sorting method for the displayed posts
Added a sorting method for the posts displayed in the main page. Added a container component that reads a part of the Redux state tree by subscribing to the store and supplies props to a presentational component it renders. Used `connect` to pass the props to the presentational component `AllPosts`.
Added a `sortOrder` property to the state tree in the Redux store to track the sort order a user wants posts to display on the screen. At the moment the sort order is hard coded, but a sorting method will be implemented sooner.