fix: Fix voting mechanism
Refactored code to fix voting mechanism. As the `AllPosts` presentational component was connected to the `VisiblePosts` container component, it needed a callback props injected as voting method. Previously the callback method was passed directly to the component via props through App.js.
Added a `votePosts` prop to the `mapDispatchToProps` function to dispatch a CHANGE_SORTING_ORDER action.