feat: Implement delete mechanism for a post
Added a `delete mechanism` to a post details View by setting a post delete property value to true.
Refactored code to refetch all posts (without their comments) and to filter the posts ids to include only the ones with the `delete` property flag set to false in the ids array passed to the reducer.