refactor: Change fields to edit in posts and comments
Changed code to edit the right fields in posts and comments, as asked by the rubric.
Fixed wrong class names and edited the CSS file.
Refactored code to update a single post, after editing it. Previously the application fetched everything from the server.
Deleted the `author` field in the edit form for both posts and comments.
Added a timestamp to comments and made it update after a comment was edited.