fix: Fix post `delete` mechanism
Fixed post `delete` mechanism by removing an action in the post reducer. Somehow code had been inadvertently added to the post reducer that returned an array of posts ids inside it (probably pasted from the `allPosts` array reducer). So, when the `post delete` button was pressed, an array of posts ids was returned in place of just updating the state of the posts object.
Added the category component selection to the `create post` and `edit post` Views.