feat: Add Modal as `add comment` button is clicked
Added a `plus button` icon to the post details view to open a Modal as a user clics it.
Added state properties to handle the modal opening and closing and to make it open and close only when the `edit modal` was closed (and viceversa).
Added an action to insert the added comment to the store state, after adding it to the server (previously all comments related to the common parent were refetched, causing a lot of work on the server side).
Replicated the same action for the post addition mechanism.