refactor: Get categories from action
Refactored code to get the categories in the reducers that need them from an action as an intermediate step to retrieve them from the server instead of having them hard coded.
Fixed code to give the `select` components where a list of posts is displayed (`AllPosts` and `Category` Views) an initial value taken from the `sortOrder` property inside the store. That value should persist when changing views.