fix: Fix `add` and `edit` mechanisms in comments form
Fixed problems in the comment forms inside the Modal.
A state value with the commentId was inhadvertently deleted in commit `97aa308...` and that resulted in the commentId not being taken anymore by the form component, which caused problems when communicating with the server.
Changed a field name in the submitted value that was addressing `comment` instead of body, resulting in an empty comment being passed to the server.